+++ egui/src/widgets/drag_value.rs
@@ -298,6 +298,11 @@ impl<'a> Widget for DragValue<'a> {
             response
         };
 
+        if response.gained_kb_focus() {
+            ui.output()
+                .push_gained_focus_event(WidgetType::DragValue, "");
+        }
+
         #[allow(clippy::float_cmp)]
         {
             response.changed = get(&mut get_set_value) != value;