+++ egui/src/widgets/selected_label.rs
@@ -55,6 +55,10 @@ impl Widget for SelectableLabel {
         let mut desired_size = total_extra + galley.size;
         desired_size.y = desired_size.y.at_least(ui.spacing().interact_size.y);
         let (rect, response) = ui.allocate_at_least(desired_size, Sense::click());
+        if response.gained_kb_focus() {
+            ui.output()
+                .push_gained_focus_event(WidgetType::SelectableLabel, &galley.text);
+        }
 
         let text_cursor = ui
             .layout()