+++ egui/src/widgets/text_edit.rs
@@ -321,6 +321,10 @@ impl<'t> TextEdit<'t> {
             Sense::hover()
         };
         let mut response = ui.interact(rect, id, sense);
+        if response.gained_kb_focus() {
+            ui.output()
+                .push_gained_focus_event(WidgetType::TextEdit, &*text);
+        }
 
         if enabled {
             if let Some(pointer_pos) = ui.input().pointer.interact_pos() {