+++ egui_demo_lib/src/apps/demo/tests.rs
@@ -93,8 +93,7 @@ impl super::Demo for ManualLayoutTest {
 
 impl super::View for ManualLayoutTest {
     fn ui(&mut self, ui: &mut egui::Ui) {
-        use egui::*;
-        reset_button(ui, self);
+        egui::reset_button(ui, self);
         let Self {
             widget_offset,
             widget_size,
@@ -107,29 +106,30 @@ impl super::View for ManualLayoutTest {
             ui.radio_value(widget_type, WidgetType::Label, "Label");
             ui.radio_value(widget_type, WidgetType::TextEdit, "TextEdit");
         });
-        Grid::new("pos_size").show(ui, |ui| {
+        egui::Grid::new("pos_size").show(ui, |ui| {
             ui.label("Widget position:");
-            ui.add(Slider::f32(&mut widget_offset.x, 0.0..=400.0));
-            ui.add(Slider::f32(&mut widget_offset.y, 0.0..=400.0));
+            ui.add(egui::Slider::f32(&mut widget_offset.x, 0.0..=400.0));
+            ui.add(egui::Slider::f32(&mut widget_offset.y, 0.0..=400.0));
             ui.end_row();
 
             ui.label("Widget size:");
-            ui.add(Slider::f32(&mut widget_size.x, 0.0..=400.0));
-            ui.add(Slider::f32(&mut widget_size.y, 0.0..=400.0));
+            ui.add(egui::Slider::f32(&mut widget_size.x, 0.0..=400.0));
+            ui.add(egui::Slider::f32(&mut widget_size.y, 0.0..=400.0));
             ui.end_row();
         });
-        let widget_rect = Rect::from_min_size(ui.min_rect().min + *widget_offset, *widget_size);
+        let widget_rect =
+            egui::Rect::from_min_size(ui.min_rect().min + *widget_offset, *widget_size);
 
         // Showing how to place a widget anywhere in the `Ui`:
         match *widget_type {
             WidgetType::Button => {
-                ui.put(widget_rect, Button::new("Example button"));
+                ui.put(widget_rect, egui::Button::new("Example button"));
             }
             WidgetType::Label => {
-                ui.put(widget_rect, Label::new("Example label"));
+                ui.put(widget_rect, egui::Label::new("Example label"));
             }
             WidgetType::TextEdit => {
-                ui.put(widget_rect, TextEdit::multiline(text_edit_contents));
+                ui.put(widget_rect, egui::TextEdit::multiline(text_edit_contents));
             }
         }
     }